Moved ThermalStorageInput from ChpTypeInput to ChpInput, thus solving bug #39 for now.
